requirements.txt
# Core dependencies
pygame==2.6.1
numpy==1.26.3
PyYAML==6.0.1
python-dotenv==1.0.0

# LLM API integration
google-generativeai==0.3.2

# Optional plotting (if needed later)
matplotlib==3.8.2

# For web-based visualization alternatives (optional)
# streamlit==1.32.0
# gradio==4.19.0